Add a 'Login' component, use axios to post to mongoDB, and implementing basic routing between Login and Register pages
I added a 'Login' component that follows the style of the 'Register' component and allows the user to login once an account has already been created.

I used axios to make POST requests and to store new user data in our database. The next step is to implement password encryption using bcrypt. So far, passwords are simply stored as strings.

Lastly, I implemented basic routing between the Login and Register pages. By default, the user lands on the Login page, but can navigate to the Register page and back.