Link our userAuth with Spotify login
I implemented a link between our userAuth and Spotify authentication. Now, whenever a user logs in, the app will prompt them to also connect their Spotify account. Information about the user's Spotify account is stored in our database; thus, if we already have their access token, their Spotify will be linked automatically (access token might have to be refreshed, the app is periodically checking for that).

If we have no record of the user's Spotify credentials in the database, the user will be prompted to connect their Spotify account and their access token will be stored in the database afterwards.

The app is keeping track whether a user is currently logged in by storing a 'username' variable in the local browser storage.