Fix auth bugs, create Profile page on the main navbar
I fixed a bug that was caused whenever a user would refresh the login page after logging in (the website would display the login form again even though that does not make sense).

I also renamed the Home page component to Profile. Home route is now considered to be the landing page. I also changed one of the redirect URIs (http://localhost:3001/spotify/callback/profile has to be added to the spotify app configuration as a new redirect URI).