Fix bug on Discover page and change the styling of search output
I fixed a bug that would cause the website to crash whenever React
would try to display the image of a certain artist, but there would
be no images of the given artist in the Spotify database. I also
changed the styling of the search output, so that it correponds to
the styling of the 'new releases'.