Fix login redirection
I modified the Login component so that automatically redirects
the user to their profile page after logging in. If they navigate
back to the login page using arrows in the browser while still
being logged in, the Login component will display what it did
before (i.e. link to the Profile page).