Various fixes
Fix bug that caused the Forgot page to crash. Style the Recover
page which was never styled before. Add an ErrorBar to profile
page that will notify the user if there are no snapshots to
retrieve from mongoDB. Make the Register button disappear from
landing page if user is logged in.