More fixes and cleanup
Found and fixed a bug that causes the Profile page to crash when
a user has no profile picture on spotify. Made the About page
responsive to changes in screen size. Removed the register button
from Discover page when a user is logged in. Got rid of unused CSS
for About page.